requirements-dev.txt
textfsm
# textfsm_aos
scrapli
scrapli[community]
scrapli[ssh2]
tox
rich
pyyaml
